requirements.txt
Click==7.0
Django==2.1.4
django-cors-headers==2.4.0
django-extensions==2.1.4
django-rq==1.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2018.7
redis==3.0.1
rq==0.12.0
six==1.11.0
